+++ Sources/TokenGeneration/Extensions/Dictionary+Extensions.swift
@@ -4,7 +4,7 @@ extension Dictionary {
     var jsonData: Data {
         get throws {
             guard let jsonData = try? JSONSerialization.data(withJSONObject: self) else {
-                throw SignedJWTGeneratorError.cantCreateJSONData
+                throw JWTGeneratorError.cantCreateJSONData
             }
             return jsonData
         }
